Cleaned up tests and renamed ReportManager to EptsReportManager: taken from esaude-reports repo
