+++ api/src/main/java/org/openmrs/module/eptsreports/api/dao/EPTSreportsDao.java
@@ -9,10 +9,12 @@
  */
 package org.openmrs.module.eptsreports.api.dao;
 
-import org.springframework.stereotype.Repository;
+import org.openmrs.api.AdministrationService;
+import org.openmrs.api.context.Context;
 import org.openmrs.api.db.hibernate.DbSession;
 import org.openmrs.api.db.hibernate.DbSessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
 
 @Repository("eptsreports.EptsReportsDao")
 public class EptsReportsDao {
@@ -20,12 +22,16 @@ public class EptsReportsDao {
 	@Autowired
 	DbSessionFactory sessionFactory;
 	
+	private AdministrationService as = Context.getAdministrationService();
+	
 	private DbSession getSession() {
 		return sessionFactory.getCurrentSession();
 	}
 	
-	public String removeReport() {
-		return "1234";
+	public void removeReportDefinition(String uuid) {
+		// String reportId = "select id from reporting_report_design where uuid='"+uuid+"'";
+		// as.executeSQL("delete from reporting_report_design_resource where report_design_id =("
+		// 		+ reportId + ");", false);
 	}
 	
 }
