+++ api/src/test/java/org/openmrs/module/eptsreports/dao/EPTSreportsDaoTest.java
@@ -25,17 +25,14 @@ import static org.junit.Assert.*;
  * rolled back by the end of each test method.
  */
 public class EptsReportsDaoTest extends BaseModuleContextSensitiveTest {
-	
+
 	@Autowired
 	EptsReportsDao dao;
-	
-	@Autowired
-	UserService userService;
-	
+
 	@Test
-	public void removeReport_shouldWork() {
-		String sut = dao.removeReport();
-		
+	public void removeReport_UuidExistsAndShouldSucceed() {
+		dao.removeReportDefinition("1234");
+
 		assertEquals(true, true);
 	}
 }
